backend/requirements.txt
asgiref==3.8.1
Brotli==1.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cssselect2==0.7.0
Django==5.1.4
django-cors-headers==4.6.0
django-extensions==3.2.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
fonttools==4.55.3
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.10.1
pyphen==0.17.0
python-decouple==3.8
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
rest-framework-simplejwt==0.0.2
setuptools==75.7.0
sqlparse==0.5.3
tinycss2==1.4.0
tinyhtml5==2.0.0
uritemplate==4.1.1
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
wheel==0.45.1
zopfli==0.2.3.post1
